Added files via upload
Changed the way of storing data into the database. Previously, all tweets are stored in the same document in couchDB, now every tweet has its own related document, which is to say, len(tweets) == len(documents).

The "search api" method has been changed from filtering by a specific query to filtering by "geocode". The format of geocode is: geocode="-37.817999,145.008244,20km" (coordinate of the centre of the circle, and the radius).