Added files via upload
This program according to the poverty rate gathering from aurin which the unit is sa2(smaller than the 9 area we suppose to),in this program I read the csv file of the 9 area respectively. And calculate the average poverty rate of all the  sub-area in the 9 area. And then use a dictionary to sort it. The key of the dictionary is the name of area and the value is the poverty of this area.