Added files via upload
This program according to the teenager rate(whose age is 15-19) gathering from aurin which the unit is sa2(smaller than the 9 area we suppose to),in this program I read the csv file of the 9 area respectively. And calculate the average teenager rate of all the  sub-area in the 9 area. And then use a dictionary to sort it. The key of the dictionary is the name of area and the value is the teenager rate of this area.